+++ flake.nix
@@ -21,13 +21,32 @@
   outputs = { self, nixpkgs, flake-utils, flake-compat, rust-overlay, pre-commit-hooks, ... }:
     flake-utils.lib.eachDefaultSystem (system:
       let
-        overlays = [ (import rust-overlay) ];
+        overlays = [ 
+          (import rust-overlay)
+          (self: super: {
+            rustc = (super.rustc.override {
+                stdenv = self.stdenv.override {
+                    targetPlatform = super.stdenv.targetPlatform // {
+                        parsed = {
+                            cpu = { name = "wasm32"; };
+                            vendor = {name = "unknown";};
+                            kernel = {name = "unknown";};
+                            abi = {name = "unknown";};
+                        };
+                    };
+                };
+            }).overrideAttrs (attrs: {
+                configureFlags = attrs.configureFlags ++ ["--set=build.docs=false"];
+            });
+          })
+        ];
         pkgs = import nixpkgs { inherit system overlays; };
         nightlyToolchain = pkgs.rust-bin.selectLatestNightlyWith
           (toolchain: toolchain.minimal.override { extensions = [ "rustfmt" ]; });
 
         stableToolchain = pkgs.rust-bin.stable.latest.minimal.override {
           extensions = [ "clippy" "llvm-tools-preview" "rust-src" ];
+          targets = ["wasm32-unknown-unknown"];
         };
       in with pkgs;
       {
@@ -62,7 +81,8 @@
             };
           };
         };
-        devShell = mkShell {
+        devShell = clang15Stdenv.mkDerivation {
+          name = "clang15-nix-shell";
           buildInputs = [
             argbash
             openssl
@@ -72,7 +92,9 @@
             stableToolchain
             nightlyToolchain
             cargo-sort
-
+            clang-tools_15
+            clangStdenv
+            llvm_15
           ] ++ lib.optionals stdenv.isDarwin [ darwin.apple_sdk.frameworks.Security ];
 
           shellHook = ''
@@ -81,6 +103,11 @@
 
             # Ensure `cargo fmt` uses `rustfmt` from nightly.
             export RUSTFMT="${nightlyToolchain}/bin/rustfmt"
+
+            export C_INCLUDE_PATH="${llvmPackages_15.libclang.lib}/lib/clang/15.0.7/include"
+            export CC=$(which clang)
+            export AR=$(which llvm-ar)
+            export CFLAGS="-mcpu=generic"
           ''
           # install pre-commit hooks
           + self.check.${system}.pre-commit-check.shellHook;