+++ primitives/src/pcs/multilinear_kzg/util.rs
@@ -12,7 +12,9 @@ use ark_poly::{
     univariate::DensePolynomial, DenseMultilinearExtension, EvaluationDomain, Evaluations,
     MultilinearExtension, Polynomial, Radix2EvaluationDomain,
 };
-use ark_std::{end_timer, format, log2, start_timer, string::ToString, sync::Arc, vec, vec::Vec};
+use ark_std::{end_timer, format, log2, start_timer, string::ToString, vec, vec::Vec};
+
+use super::MLE;
 
 /// Evaluate eq polynomial. use the public one later
 #[cfg(any(test, feature = "test-srs"))]
@@ -130,7 +132,7 @@ pub fn get_batched_nv(num_var: usize, polynomials_len: usize) -> usize {
 /// merge a set of polynomials. Returns an error if the
 /// polynomials do not share a same number of nvs.
 pub fn merge_polynomials<F: PrimeField>(
-    polynomials: &[Arc<DenseMultilinearExtension<F>>],
+    polynomials: &[MLE<F>],
 ) -> Result<DenseMultilinearExtension<F>, PCSError> {
     let nv = polynomials[0].num_vars();
     for poly in polynomials.iter() {
@@ -192,7 +194,7 @@ pub(crate) fn build_l<F: PrimeField>(
 // are included in the `batch_proof`.
 #[cfg(test)]
 pub(crate) fn generate_evaluations<F: PrimeField>(
-    polynomials: &[Arc<DenseMultilinearExtension<F>>],
+    polynomials: &[MLE<F>],
     points: &[Vec<F>],
 ) -> Result<Vec<F>, PCSError> {
     if polynomials.len() != points.len() {
@@ -316,7 +318,7 @@ mod test {
         // 1, 0 |-> 0
         // 1, 1 |-> 5
         let w_eval = vec![F::zero(), F::from(2u64), F::zero(), F::from(5u64)];
-        let w1 = Arc::new(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
+        let w1 = MLE::from(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
 
         // W2 = x1x2 + x1 whose evaluations are
         // 0, 0 |-> 0
@@ -324,7 +326,7 @@ mod test {
         // 1, 0 |-> 1
         // 1, 1 |-> 2
         let w_eval = vec![F::zero(), F::zero(), F::from(1u64), F::from(2u64)];
-        let w2 = Arc::new(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
+        let w2 = MLE::from(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
 
         // W3 = x1 + x2 whose evaluations are
         // 0, 0 |-> 0
@@ -332,7 +334,7 @@ mod test {
         // 1, 0 |-> 1
         // 1, 1 |-> 2
         let w_eval = vec![F::zero(), F::one(), F::from(1u64), F::from(2u64)];
-        let w3 = Arc::new(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
+        let w3 = MLE::from(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
 
         {
             // W = (3x1x2 + 2x2)(1-x0) + (x1x2 + x1)x0
@@ -565,15 +567,15 @@ mod test {
         // Example from page 53:
         // W1 = 3x1x2 + 2x2
         let w_eval = vec![Fr::zero(), Fr::from(2u64), Fr::zero(), Fr::from(5u64)];
-        let w1 = Arc::new(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
+        let w1 = MLE::from(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
 
         // W2 = x1x2 + x1
         let w_eval = vec![Fr::zero(), Fr::zero(), Fr::from(1u64), Fr::from(2u64)];
-        let w2 = Arc::new(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
+        let w2 = MLE::from(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
 
         // W3 = x1 + x2
         let w_eval = vec![Fr::zero(), Fr::one(), Fr::from(1u64), Fr::from(2u64)];
-        let w3 = Arc::new(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
+        let w3 = MLE::from(DenseMultilinearExtension::from_evaluations_vec(2, w_eval));
 
         let r = Fr::from(42u64);